Request content validation, invalid API Request error handling

+++ app/controllers/todo.controller.js
@@ -1,8 +1,15 @@
 const todoService = require('../services/todo.service');
-const validator = require('../models/todo.validator');
+// model, model validation
+const modelValidator = require('../validators/todo.validator');
 const TodoItem = require('../models/todo.model');
+// request validation
+const requestValidator = require('../validators/request.validator');
 
 async function getTodoList(req, res, next){
+    const reqNotValidated = requestValidator(req);
+    if (reqNotValidated) {
+        return next(reqNotValidated);
+    }
     try {
         const returnFromService = await todoService.getTodoList();
         return res.status(200).send(returnFromService);
@@ -12,8 +19,12 @@ async function getTodoList(req, res, next){
 }
 
 async function addTodoItem(req, res, next){
+    const reqNotValidated = requestValidator(req);
+    if (reqNotValidated) {
+        return next(reqNotValidated);
+    }
     try {
-        const returnFromService = await validator(todoService.addTodoItem(req, res, next));
+        const returnFromService = await modelValidator(todoService.addTodoItem(req, res, next));
         const todoItem = new TodoItem(returnFromService);
         const newItem = await todoItem.save()
         return res.status(200).send(newItem);
@@ -23,6 +34,10 @@ async function addTodoItem(req, res, next){
 }
 
 async function deleteTodoItem(req, res, next){
+    const reqNotValidated = requestValidator(req);
+    if (reqNotValidated) {
+        return next(reqNotValidated);
+    }
     try {
         const returnFromService = todoService.deleteTodoItem(req);
         const todoItem = await TodoItem.findById(returnFromService);
@@ -34,12 +49,19 @@ async function deleteTodoItem(req, res, next){
 }
 
 async function changeTodoItem(req, res, next){
+    const reqNotValidated = requestValidator(req);
+    if (reqNotValidated) {
+        return next(reqNotValidated);
+    }
     try {
         const returnFromService = await todoService.changeTodoItem(req);
         const { id, done } = returnFromService;
         const updatedItem = await TodoItem.findByIdAndUpdate(id, { "done": done }, { new: true });
         return res.status(200).send(updatedItem);
     } catch(err) {
+        if (err.message === "Cannot read property 'done' of null") {
+            err.message = "There is no such to-do item in list ;("
+        }
         next(err);
     }
 }
